Extract code that computes dependencies.
